Extract ParSpatialMatcherUtils
