+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/matcher/SpatialMatcher.scala
@@ -15,6 +15,7 @@ import coop.rchain.models.rholang.implicits.{VectorPar, _}
 import coop.rchain.rholang.interpreter.accounting.{Cost, _}
 import coop.rchain.rholang.interpreter.errors.OutOfPhlogistonsError
 import coop.rchain.rholang.interpreter.matcher.NonDetFreeMapWithCost._
+import coop.rchain.rholang.interpreter.matcher.ParSpatialMatcherUtils.{noFrees, subPars}
 import coop.rchain.rholang.interpreter.matcher.SpatialMatcher._
 import coop.rchain.rholang.interpreter.matcher.StreamT._
 
@@ -79,127 +80,6 @@ object SpatialMatcher extends SpatialMatcherInstances {
       override def spatialMatch(target: T, pattern: P): NonDetFreeMapWithCost[Unit] =
         fn(target, pattern)
     }
-
-  def noFrees(par: Par): Par =
-    par.withExprs(noFrees(par.exprs))
-
-  def noFrees(exprs: Seq[Expr]): Seq[Expr] =
-    exprs.filter({ (expr) =>
-      expr.exprInstance match {
-        case EVarBody(EVar(v)) =>
-          v.varInstance match {
-            case FreeVar(_)  => false
-            case Wildcard(_) => false
-            case _           => true
-          }
-        case _ => true
-      }
-    })
-
-  def subPars(
-      par: Par,
-      min: ParCount,
-      max: ParCount,
-      minPrune: ParCount,
-      maxPrune: ParCount
-  ): Stream[(Par, Par)] = {
-
-    val sendMax    = math.min(max.sends, par.sends.size - minPrune.sends)
-    val receiveMax = math.min(max.receives, par.receives.size - minPrune.receives)
-    val newMax     = math.min(max.news, par.news.size - minPrune.news)
-    val exprMax    = math.min(max.exprs, par.exprs.size - minPrune.exprs)
-    val matchMax   = math.min(max.matches, par.matches.size - minPrune.matches)
-    val idMax      = math.min(max.ids, par.ids.size - minPrune.ids)
-    val bundleMax  = math.min(max.bundles, par.bundles.size - minPrune.bundles)
-
-    val sendMin    = math.max(min.sends, par.sends.size - maxPrune.sends)
-    val receiveMin = math.max(min.receives, par.receives.size - maxPrune.receives)
-    val newMin     = math.max(min.news, par.news.size - maxPrune.news)
-    val exprMin    = math.max(min.exprs, par.exprs.size - maxPrune.exprs)
-    val matchMin   = math.max(min.matches, par.matches.size - maxPrune.matches)
-    val idMin      = math.max(min.ids, par.ids.size - maxPrune.ids)
-    val bundleMin  = math.max(min.bundles, par.bundles.size - maxPrune.bundles)
-
-    for {
-      subSends    <- minMaxSubsets(par.sends, sendMin, sendMax)
-      subReceives <- minMaxSubsets(par.receives, receiveMin, receiveMax)
-      subNews     <- minMaxSubsets(par.news, newMin, newMax)
-      subExprs    <- minMaxSubsets(par.exprs, exprMin, exprMax)
-      subMatches  <- minMaxSubsets(par.matches, matchMin, matchMax)
-      subIds      <- minMaxSubsets(par.ids, idMin, idMax)
-      subBundles  <- minMaxSubsets(par.bundles, bundleMin, bundleMax)
-    } yield
-      (
-        Par(
-          subSends._1,
-          subReceives._1,
-          subNews._1,
-          subExprs._1,
-          subMatches._1,
-          subIds._1,
-          subBundles._1
-        ),
-        Par(
-          subSends._2,
-          subReceives._2,
-          subNews._2,
-          subExprs._2,
-          subMatches._2,
-          subIds._2,
-          subBundles._2
-        )
-      )
-  }
-
-  def minMaxSubsets[A](as: Seq[A], minSize: Int, maxSize: Int): Stream[(Seq[A], Seq[A])] = {
-    def countedMaxSubsets(as: Seq[A], maxSize: Int): Stream[(Seq[A], Seq[A], Int)] =
-      as match {
-        case Nil => Stream((as, as, 0))
-        case head +: rem =>
-          (as.slice(0, 0), as, 0) #::
-            (for {
-              countedTail               <- countedMaxSubsets(rem, maxSize)
-              (tail, complement, count) = countedTail
-              result <- {
-                if (count == maxSize)
-                  Stream((tail, head +: complement, count))
-                else if (tail.isEmpty)
-                  Stream((head +: tail, complement, 1))
-                else
-                  Stream((tail, head +: complement, count), (head +: tail, complement, count + 1))
-              }
-            } yield result)
-      }
-    def worker(as: Seq[A], minSize: Int, maxSize: Int): Stream[(Seq[A], Seq[A], Int)] =
-      if (maxSize < 0)
-        Stream.empty
-      else if (minSize > maxSize)
-        Stream.empty
-      else if (minSize <= 0)
-        if (maxSize == 0)
-          Stream((as.slice(0, 0), as, 0))
-        else
-          countedMaxSubsets(as, maxSize)
-      else
-        as match {
-          case Nil => Stream.empty
-          case hd +: rem =>
-            val decr = minSize - 1
-            for {
-              countedTail               <- worker(rem, decr, maxSize)
-              (tail, complement, count) = countedTail
-              result <- {
-                if (count == maxSize)
-                  Stream((tail, hd +: complement, count))
-                else if (count == decr)
-                  Stream((hd +: tail, complement, minSize))
-                else
-                  Stream((tail, hd +: complement, count), (hd +: tail, complement, count + 1))
-              }
-            } yield result
-        }
-    worker(as, minSize, maxSize).map(x => (x._1, x._2))
-  }
 
   private[matcher] type F[A] = NonDetFreeMapWithCost[A]
   private[matcher] val F = NonDetFreeMapWithCost
